requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
idna==2.10
lxml==4.6.2
Pillow==8.2.0
PyMySQL==1.0.2
pytz==2021.1
requests==2.25.1
simplejson==3.17.2
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
tushare==1.2.62
urllib3==1.26.4
websocket-client==0.58.0
